Ataualização

+++ Models/Produto.cs
@@ -11,7 +11,7 @@ namespace Estoque.Models
         public int ProdutoId { get; set; }
         public string Nome { get; set; }
         public int Quantidade { get; set; }
-        public int Preco { get; set; }
+        public double Preco { get; set; }
         public int CategoriaId { get; set; }
         public Categoria Categoria { get; set; }
